+++ src/load/HISTORY.rst
@@ -4,7 +4,7 @@ Release History
 ===============
 1.5.0
 ++++++
-* Add support for engine reference identity using CLI. Engine reference identity can be set using `--engine-reference-identity-type` and `--engine-reference-identities` argument in 'az load test create' and 'az load test update' commands. Engine reference identity set in YAML config file under key `referenceIdentities` with `kind` as `Engine` will also be honoured.
+* Add support for engine reference identity using CLI. Engine reference identity can be set using `--engine-ref-id-type` and `--engine-ref-ids` argument in 'az load test create' and 'az load test update' commands. Engine reference identity set in YAML config file under key `referenceIdentities` with `kind` as `Engine` will also be honoured.
 
 1.4.3
 ++++++